Playerクラスのファイル作成
